requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter
torch-sparse
torch-cluster
pykeen>=1.10.0
networkx>=3.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
rdkit>=2023.3.1
biopython>=1.81
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0
pyyaml>=6.0
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
pyarrow>=12.0.0
